requirements.txt
# Core AI libraries - Use compatible versions
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2

# Data processing
pandas==2.0.3
numpy==1.24.3

# Vector databases
chromadb==0.4.18
faiss-cpu==1.7.4

# Token counting
tiktoken==0.5.1

# Web framework (optional)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing and evaluation
pytest==7.4.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.1
seaborn==0.13.0

# HTTP client
httpx==0.25.0